在原bv的基础上做了一些修改： - 主屏（整体框架以及首页、UGC、PGC）     - 重构左侧导航栏，移除抽屉展开效果     - 首页、ugc：标题改为2行、缩减视频列表的间距     - 首页、ugc：视频卡片显示发布时间     - 移除 ugc、pgc 页面的轮播组件，避免抢占焦点导致内容区tab、左侧导航 操作不畅     - 记住每个内容页当前选中的Tab并在切换回来后恢复     - 支持按返回键回到左侧菜单栏     - 首页、ugc、pgc：顶部Tab切换增加防抖，延迟发起内容和请求     - 首页、ugc、pgc：简化内容区切换动画；让容器填充满屏幕，避免竖向的收缩动画     - 首页、ugc、pgc：按返回键定位到顶部tab时，内容区不滚动到顶部（点击tab刷新数据会回到顶部）     - UGC：缓存每个tab的数据，减少请求次数     - 左侧菜单项未获得焦点的时候，去掉选中效果     - 左侧菜单切换增加防抖，延迟切换右侧内容。 - 播放详情页     - 增加点赞、投币     - 调整收藏、简介入口位置     - 修改视频封面，合集中的视频改成显示视频封面。原逻辑合集中的视频显示合集的封面     - 选中封面增加边框     - 多个收藏夹时，弹窗让用户选择要添加到哪个收藏夹，仅单个收藏夹时不弹窗直接加入默认收藏夹     - 视频详情页增加实例管理逻辑，最多保留3个实例，优化内存占用，同时也更少返回次数就能回到列表页     - 回退页面的时不请求视频关联的用户数据（点赞、收藏），减少请求时有必要的 - 播放页     - 调小标题字体、调小进度条高度、调浅缓冲进度颜色     - 播放速度调成"画面音频"子菜单的第一个     - 播放速度生效周期改成单个视频，即切视频就重置为1（原版播放速度是全局存储的）     - 增加按下返回键隐藏控制条     - 增加快进/快退一段时间无操作自动确认播放     - 支持“竖屏视频播放时的最大清晰度为1080P”（解决部分设备竖屏视频变形的问题）     - 快退一次从5s改为10s，和快进保持一致     - 支持“都播完后退出播放器”，退出后是回到视频详情页     - 支持隐藏视频播放页面左下角的视频调试信息     - 优化视频缓冲逻辑，缓解卡面卡死 - 搜索     - 搜索结果页视频卡片显示发布时间     - 支持按返回键回到左侧菜单栏     - 搜索结果页改成选中tab的时候才发起请求，以便减少请求次数 - up空间页     - 视频卡片显示发布时间 - 收藏     - 视频卡片显示 收藏时间 - 浏览历史     - 视频卡片显示 浏览时间 - 标签搜索结果页     - 视频卡片显示发布时间 - 设置页面     - 更多设置-增加设置“设置竖屏视频播放时的最大清晰度为1080P”，默认禁用     - 更多设置-增加设置“都播完后退出播放器”，默认开启     - 界面设置-增加设置“播放器显示视频调试信息”，默认关闭

+++ bili-api/src/main/kotlin/dev/aaa1115910/biliapi/http/entity/history/HistoryData.kt
@@ -90,7 +90,7 @@ data class HistoryItem(
     @SerialName("author_mid")
     val authorMid: Long,
     @SerialName("view_at")
-    val viewAt: Int,
+    val viewAt: Long,
     val progress: Int,
     val badge: String,
     @SerialName("show_title")
